chore: convert storybook configs to typescript
