--- .storybook/main.js
@@ -1,8 +1,0 @@
-module.exports = {
-  "stories": ["../src/**/*.stories.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
-  "addons": ["@storybook/addon-links", "@storybook/addon-essentials", "@storybook/addon-interactions"],
-  "framework": {
-    name: "@storybook/react-webpack5",
-    options: {}
-  }
-};
