+++ .storybook/main.ts
@@ -1,0 +1,15 @@
+import type { StorybookConfig } from '@storybook/react-webpack5';
+
+const config: StorybookConfig = {
+  "stories": [
+    "../src/**/*.mdx",
+    "../src/**/*.story.@(js|jsx|ts|tsx)"
+  ],
+  "addons": ["@storybook/addon-links", "@storybook/addon-essentials", "@storybook/addon-interactions"],
+  "framework": {
+    name: "@storybook/react-webpack5",
+    options: {}
+  }
+};
+
+export default config;
